requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
pydantic
paddleocr
httpx
ollama 